+++ lib/papa/larga.rb
@@ -5,16 +5,14 @@ module Papa
     RELEASE_OR_HOTFIX_PROTECTION = 'off'
     DEFAULT_PROTECTION = 'on'
 
+    def self.type
+      Larga::Type.new
+    end
+
     def self.deploy(branch:, lifespan: DEFAULT_LIFESPAN, protection: DEFAULT_PROTECTION, hostname: nil)
-      options = []
-      options << '-action deploy',
-      options << "-branch #{branch}",
-      options << "-lifespan #{lifespan}",
-      options << "-protection #{protection}",
-      if hostname
-        options << "-hostname #{hostname}"
-      end
-      Command.new "larga #{options.join(' ')}"
+      require 'papa/larga/deploy'
+
+      Larga::Deploy.new(branch, lifespan, protection, hostname)
     end
   end
 end